pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "baby_care_ai"
version = "0.1.0"
description = "A system for collecting and syncing baby care images"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp==3.13.2",
    "blinkpy==0.25.2",
    "python-dotenv==1.2.1",
    "Pillow==12.0.0",
    "ImageHash==4.3.2",
    "PyDrive2==1.21.3",
]

[project.scripts]
baby-care-automation = "baby_care_ai.automation_logic:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["baby_care_ai*"]
